Consolidated whole genome controls rendering code to controls script; resolved integrations test failure
